AudioFrame: split hpp/cpp
